requirements.txt
pandas~=2.1.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
python-jenkins~=1.8.1
Flask~=2.0.3
Flask-Cors~=3.0.10
atlassian-python-api~=3.41.2
Jinja2~=3.0.3
requests~=2.27.1
urllib3~=1.26.9
Werkzeug~=2.0.3
xlrd~=2.0.1
xlwt~=1.3.0
lxml~=4.9.3
PyMySQL~=0.10.1
allure-pytest~=2.13.2
allure-python-commons~=2.13.2
selenium~=4.12.0
sortedcontainers~=2.4.0
openai~=0.28.0
crypto~=1.4.1